Move emitter out to own file to include in all units
