+++ ts-events/emitter.ts
@@ -1,0 +1,4 @@
+import events = require("events")
+
+// @klotho::pubsub
+export const MyEmitter = new events.EventEmitter();
