+++ ts-events/index.ts
@@ -4,11 +4,9 @@
  * }
  */
 
-import events = require("events")
 import * as Express from "express"
+import { MyEmitter } from "./emitter"
 
-// @klotho::pubsub
-export const MyEmitter = new events.EventEmitter();
 
 import "./hello_handler" // import after emitter set up to register handlers
 import "./other/handler"
